// Angular-Project/src/app/models/User.ts
export interface User {
  exists?: boolean;
  id?: string;
  address?: string;
  area?: string;
  balance?: number;
  email?: string;
  location?: string;
  name?: string;
  orderConfirm?: Array<string>;
  orderDate?: string;
  orderHard?: Array<string>;
  orderId?: string;
  orderImage?: any;
  orderTime?: string;
  phone?: string;
  userPhoto?: any;

}